requirements.txt
flask
onnxruntime
torchvision
Pillow
flask-cors